Allow user supplied string:string data to be stored in the JWT (#12)
This closes #10 by allowing users to supply a `TokenData` `[string]string` map that will be embedded into the JWT on a `GetToken` call. The decrypted and parsed TokenData is available in the `*Claims` struct that the user gets back from a call to `Validate`.

In the future, it may make sense to allow the TokenData to map `[string]interface{}` but for now it's simpler and more opinionated to force the caller to serialize to a string.